+++ src/pipeline/aodv_pipeline.h
@@ -111,42 +111,42 @@ struct aodv_msg_hello {
 } __attribute__((__packed__));
 
 // ------------- pipeline -----------------------------------------------------
-int aodv_handle_hello(dessert_msg_t* msg, size_t len,
+int aodv_handle_hello(dessert_msg_t* msg, uint32_t len,
                       dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_handle_rreq(dessert_msg_t* msg, size_t len,
+int aodv_handle_rreq(dessert_msg_t* msg, uint32_t len,
                      dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_handle_rerr(dessert_msg_t* msg, size_t len,
+int aodv_handle_rerr(dessert_msg_t* msg, uint32_t len,
                      dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_handle_rrep(dessert_msg_t* msg, size_t len,
+int aodv_handle_rrep(dessert_msg_t* msg, uint32_t len,
                      dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_forward_broadcast(dessert_msg_t* msg, size_t len,
+int aodv_forward_broadcast(dessert_msg_t* msg, uint32_t len,
                            dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_forward_multicast(dessert_msg_t* msg, size_t len,
+int aodv_forward_multicast(dessert_msg_t* msg, uint32_t len,
                            dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
-int aodv_forward(dessert_msg_t* msg, size_t len,
+int aodv_forward(dessert_msg_t* msg, uint32_t len,
                  dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
 /**
  * Encapsulate packets as dessert_msg,
  * set NEXT HOP if known and send via AODV routing protocol
  */
-int aodv_sys2rp(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc,
+int aodv_sys2rp(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc,
                 dessert_sysif_t* sysif, dessert_frameid_t id);
 
-int aodv_sys_drop_multicast(dessert_msg_t* msg, size_t len, dessert_msg_proc_t* proc,
+int aodv_sys_drop_multicast(dessert_msg_t* msg, uint32_t len, dessert_msg_proc_t* proc,
                             dessert_sysif_t* sysif, dessert_frameid_t id);
 /** forward packets received via AODV to tun interface */
-int aodv_local_unicast(dessert_msg_t* msg, size_t len,
+int aodv_local_unicast(dessert_msg_t* msg, uint32_t len,
                        dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
 /** drop errors (drop corrupt packets, packets from myself and etc...)*/
-int aodv_drop_errors(dessert_msg_t* msg, size_t len,
+int aodv_drop_errors(dessert_msg_t* msg, uint32_t len,
                      dessert_msg_proc_t* proc, dessert_meshif_t* iface, dessert_frameid_t id);
 
 // ------------------------------ periodic ----------------------------------------------------
